cantinero-scraper/requirements.txt
fastapi
uvicorn
scrapy
mssql-python
python-dotenv
requests